+++ deltafi-action-kit/src/test/java/org/deltafi/actionkit/action/util/ActionParameterSchemaGeneratorTest.java
@@ -28,7 +28,6 @@ import java.util.Objects;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ActionParameterSchemaGeneratorTest {
-
     @Test
     void testGetSchema() throws IOException {
         JsonNode schemaJson = ActionParameterSchemaGenerator.generateSchema(TestActionParameters.class);