+++ deltafi-core/src/main/java/org/deltafi/core/services/IngressFlowService.java
@@ -17,6 +17,7 @@
  */
 package org.deltafi.core.services;
 
+import lombok.extern.slf4j.Slf4j;
 import org.deltafi.core.converters.IngressFlowPlanConverter;
 import org.deltafi.core.repo.IngressFlowRepo;
 import org.deltafi.core.snapshot.SystemSnapshot;
@@ -26,8 +27,11 @@ import org.deltafi.core.validation.IngressFlowValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @Service
+@Slf4j
 public class IngressFlowService extends FlowService<IngressFlowPlan, IngressFlow> {
 
     private static final IngressFlowPlanConverter INGRESS_FLOW_PLAN_CONVERTER = new IngressFlowPlanConverter();
@@ -49,8 +53,47 @@ public class IngressFlowService extends FlowService<IngressFlowPlan, IngressFlow
     }
 
     @Override
-    List<String> getTestModeFromSnapshot( SystemSnapshot systemSnapshot) {
+    List<String> getTestModeFromSnapshot(SystemSnapshot systemSnapshot) {
         return systemSnapshot.getTestIngressFlows();
     }
 
+    /**
+     * Sets the maximum number of errors allowed for a given flow, identified by its name.
+     * If the maximum errors for the flow are already set to the specified value, the method
+     * logs a warning and returns false. If the update is successful, the method refreshes the
+     * cache and returns true.
+     *
+     * @param flowName The name of the flow to update, represented as a {@code String}.
+     * @param maxErrors The new maximum number of errors to be set for the specified flow, as an {@code int}.
+     * @return A {@code boolean} value indicating whether the update was successful (true) or not (false).
+     */
+    public boolean setMaxErrors(String flowName, int maxErrors) {
+        IngressFlow flow = getFlowOrThrow(flowName);
+
+        if (flow.getMaxErrors() == maxErrors) {
+            log.warn("Tried to set max errors on ingress flow {} to {} when already set", flowName, maxErrors);
+            return false;
+        }
+
+        if (((IngressFlowRepo) flowRepo).updateMaxErrors(flowName, maxErrors)) {
+            refreshCache();
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * Retrieves a map containing the maximum number of errors allowed per flow.
+     * This method filters out flows with a maximum error count of 0, only including
+     * those with a positive maximum error count.
+     *
+     * @return A {@code Map<String, Integer>} where each key represents a flow name,
+     * and the corresponding value is the maximum number of errors allowed for that flow.
+     */
+    public Map<String, Integer> maxErrorsPerFlow() {
+        return flowCache.entrySet().stream()
+                .filter(e -> e.getValue().getMaxErrors() > 0)
+                .collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().getMaxErrors()));
+    }
 }