+++ deltafi-core/src/test/java/org/deltafi/core/services/IngressServiceTest.java
@@ -69,6 +69,10 @@ class IngressServiceTest {
     @Mock
     FlowAssignmentService flowAssignmentService;
 
+    @Mock
+    @SuppressWarnings("unused")
+    ErrorCountService errorCountService;
+
     @Spy
     DeltaFiPropertiesService deltaFiPropertiesService = new MockDeltaFiPropertiesService();